updated models to use decimal instead of integer for locations

+++ server/models/hazards.js
@@ -30,23 +30,23 @@ Hazards.init(
 
     // location
     accuracy: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(5, 18),
     },
     altitude: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(5, 18),
     },
     altitudeAccuracy: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(5, 18),
     },
     heading: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(5, 18),
     },
     latitude: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(5, 18),
       allowNull: false,
     },
     longitude: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(5, 18),
       allowNull: false,
     },
   },
