+++ server/models/trailPoints.js
@@ -6,27 +6,27 @@ class TrailPoints extends Model {}
 TrailPoints.init(
   {
     accuracy: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(5, 18),
     },
     altitude: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(5, 18),
     },
     altitudeAccuracy: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(5, 18),
     },
     heading: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(5, 18),
     },
     latitude: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(5, 18),
       allowNull: false,
     },
     longitude: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(5, 18),
       allowNull: false,
     },
     speed: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(5, 18),
     },
     timestamp: {
       type: DataTypes.DATE,
